+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 This changelog starts with the restructured 1.0.0 release that happened after context isolation changes. The changelogs here should more-or-less mirror the ones that get shown in the client but probably with less formatting and pizzazz.
 
+## 1.5.1
+
+### Added
+- `Number` settings component using builtin `input[type=number]`
+- Settings for monaco editor: font size, line numbers, minimap, whitespace, quick suggestions, tooltips
+
+### Removed
+
+### Changed
+
+### Fixed
+- Fixed `GuildComponent` throwing errors
+- Fixed public servers button disappearing when guilds list rerenders
+- Fixed plugin compilation error not pointing to console
+- Fixed plugins with no `@name` attempting to load
+
 ## 1.5.0
 
 ### Added